Fifth task

+++ src/AbsoluteJava.java
@@ -5,7 +5,7 @@ public class AbsoluteJava {
 
         Scanner input = new Scanner(System.in);
         System.out.println("Insert a number:");
-        int a = input.nextInt();
+        double a = input.nextDouble();
         System.out.println("Absolute value of inserted number is: " + Math.abs(a));
 
     }
